+++ product-service/package.json
