+++ product-service/src/functions/create-product/handler.ts
