+++ product-service/src/functions/create-product/index.ts
@@ -12,3 +12,4 @@ export default {
     },
   ],
 };
+
