+++ product-service/src/functions/get-product-by-id/handler.ts
