+++ product-service/src/functions/get-product-by-id/index.ts
