+++ product-service/src/functions/get-products-list/handler.ts
