+++ product-service/src/libs/dynamo-db-connect-client.ts
