+++ product-service/src/libs/handler-resolver.ts
