+++ product-service/src/libs/lambda.ts
