+++ product-service/src/mocks/index.ts
