+++ product-service/src/mocks/products.json
