+++ product-service/src/mocks/products.ts
