+++ product-service/src/mocks/stocks.json
