+++ product-service/src/repositories/stock.repository.ts
