+++ product-service/src/schemas/index.ts
