+++ product-service/src/schemas/product.schema.ts
