+++ product-service/src/schemas/stock.schema.ts
