+++ product-service/tsconfig.json
